add BVH tree accelerate ray tracing

+++ RayTracing/Head/Camera.h
@@ -38,7 +38,7 @@ public:
 	DOFCamera(vec3 lookfrom, vec3 lookat, vec3 vup, float vfov, float aspect, float aperture, float focus_dist) {
 		// vfov is top to bottom in degrees
 		lens_radius = aperture / 2;
-		float theta = vfov * M_PI / 180;
+		float theta = vfov * (float)M_PI / 180;
 		float half_height = tan(theta / 2)* focus_dist;
 		float half_width = aspect * half_height;
 		origin = lookfrom;
@@ -73,7 +73,7 @@ public:
 		time0 = t0;
 		time1 = t1;
 		lens_radius = aperture / 2;
-		float theta = vfov * M_PI / 180;
+		float theta = vfov * (float)M_PI / 180;
 		float half_height = tan(theta / 2);
 		float half_width = aspect * half_height;
 		origin = lookfrom;
