+++ RayTracing/Head/Material.h
@@ -8,14 +8,14 @@ struct Hit_Record;
 float schlick(float cosine, float ref_idx) {
 	float r0 = (1 - ref_idx) / (1 + ref_idx);
 	r0 = r0 * r0;
-	return r0 + (1 - (double)r0) * pow((1 - cosine), 5);
+	return r0 + (1 - r0) * (float)pow((1 - cosine), 5);
 }
 
 
 bool refract(const vec3& v, const vec3& n, float ni_over_nt, vec3& refracted) {
 	vec3 uv = unit_vector(v);
 	float dt = dot(uv, n);
-	float discriminant = 1.0f - (double)ni_over_nt * ni_over_nt * (1 - (double)dt * dt);
+	float discriminant = 1.0f - ni_over_nt * ni_over_nt * (1 - dt * dt);
 	if (discriminant > 0) {
 		refracted = ni_over_nt * (uv - n * dt) - n * sqrt(discriminant);
 		return true;
@@ -93,7 +93,7 @@ public:
 		}
 		else {
 			outward_normal = rec.normal;
-			ni_over_nt = 1.0 / ref_idx;
+			ni_over_nt = 1.0f / ref_idx;
 			cosine = -dot(r_in.direction(), rec.normal) / r_in.direction().length();
 		}
 		if (refract(r_in.direction(), outward_normal, ni_over_nt, refracted))
